refactor(scripts): unify barrel-file manipulation into one module
Three independent index.ts re-export implementations existed: setup-project
(removal), bundle-installer (find/insert), and generate-component (append).
Extract the pure string transforms into lib/scripts/barrel-file.ts
(findBarrelLine, removeBarrelLines, insertBarrelLine); all three callers now
delegate. Each caller keeps its own I/O (dry-run, logging, create-if-missing).

Net -83 lines across callers; single source of truth for barrel matching and
sorted idempotent insertion.
